# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Basic CRUD Laravel API Documentation",
        "description": "API Documentation - Basic CRUD Laravel",
        "contact": {
            "email": "info@app-appraisal.com"
        },
        "license": {
            "name": "@app-appraisal",
            "url": "https://app-appraisal.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/contact-companies/{id}": {
            "get": {
                "tags": [
                    "Contact Companies"
                ],
                "summary": "Get Contact Company detail",
                "operationId": "show",
                "responses": {
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contact Companies"
                ],
                "summary": "Update an existing Contact Company",
                "operationId": "update",
                "responses": {
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contact Companies"
                ],
                "summary": "Delete a Contact Company",
                "operationId": "destroy",
                "responses": {
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/contact-companies/{id}/restore": {
            "post": {
                "tags": [
                    "Contact Companies"
                ],
                "summary": "Restore a Contact Company from trash",
                "operationId": "restore",
                "responses": {
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "JWT",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}